feat: 提取 authority 至单独的 依赖中，仅仅保留权限相关逻辑，提供相关接口，authority starter 依赖 authority 并处理web中的一些事项

+++ boot/authority-web-boot-starter/src/main/java/com/kewen/framework/boot/authority/advance/authcheck/ApplicationBusiness.java
@@ -10,5 +10,5 @@ public interface ApplicationBusiness {
      * 获取到关联的ID
      * @return 业务关联ID，业务具体事项的主键
      */
-    Integer getBusinessId();
+    Long getBusinessId();
 }
