+++ boot/authority-web-boot-starter/src/main/java/com/kewen/framework/boot/authority/advance/authcheck/DefaultApplicationBusiness.java
@@ -10,9 +10,9 @@ import lombok.Data;
 @Data
 public class DefaultApplicationBusiness implements ApplicationBusiness {
 
-    private Integer businessId;
+    private Long businessId;
     @Override
-    public Integer getBusinessId() {
+    public Long getBusinessId() {
         return businessId;
     }
 }
